+++ tdms/src/tests.rs
@@ -1,0 +1,89 @@
+use tdms_format::segment::LeadIn;
+
+use crate::segment::Segment;
+use crate::TDMSFile;
+use std::fs::File;
+use std::io::Read;
+use std::path::Path;
+
+#[test]
+fn can_read_lead_in() {
+    let f = File::open(Path::new("data/standard.tdms")).expect("Failure to open file");
+    let mut buffer: Vec<u8> = Vec::with_capacity(LeadIn::SIZE);
+    f.take(LeadIn::SIZE as u64)
+        .read_to_end(&mut buffer)
+        .unwrap();
+    let lead_in: LeadIn = match LeadIn::from_bytes(&mut buffer) {
+        Ok(s) => s,
+        Err(e) => panic!("{:?}", e),
+    };
+
+    assert_eq!(lead_in.tag, [84, 68, 83, 109]);
+    assert_eq!(lead_in.version_number, 4713);
+    assert_eq!(lead_in.next_segment_offset, 292862);
+    assert_eq!(lead_in.raw_data_offset, 4862);
+}
+
+#[test]
+fn can_read_all_() {
+    let file = match TDMSFile::from_path(Path::new("data/standard.tdms")) {
+        Ok(f) => f,
+        Err(e) => panic!("{:?}", e),
+    };
+
+    assert_eq!(file.segments.len(), 2);
+}
+
+#[test]
+fn can_read_all_groups_from_segment() {
+    let file = match TDMSFile::from_path(Path::new("data/standard.tdms")) {
+        Ok(f) => f,
+        Err(e) => panic!("{:?}", e),
+    };
+
+    assert_eq!(file.segments.len(), 2);
+}
+
+#[test]
+fn can_read_all_segments_be() {
+    let file = match TDMSFile::from_path(Path::new("data/big_endian.tdms")) {
+        Ok(f) => f,
+        Err(e) => panic!("{:?}", e),
+    };
+
+    assert_eq!(file.segments.len(), 2);
+}
+
+#[test]
+fn can_read_all_segments_raw() {
+    let file = match TDMSFile::from_path(Path::new("data/raw.tdms")) {
+        Ok(f) => f,
+        Err(e) => panic!("{:?}", e),
+    };
+
+    assert_eq!(file.segments.len(), 3);
+}
+
+#[test]
+fn can_read_groups_channels() {
+    let file = match TDMSFile::from_path(Path::new("data/standard.tdms")) {
+        Ok(f) => f,
+        Err(e) => panic!("{:?}", e),
+    };
+
+    let groups = file.groups();
+    assert!(groups.len() > 0);
+
+    /*    for group in groups {
+        let channels = file.channels(&group);
+
+        for (channel, data_type) in channels {
+            let full_channel = match file.channel_double_float(&group, &channel) {
+                Ok(c) => c,
+                Err(e) => panic!("{:?}", e),
+            };
+        }
+    }*/
+
+    assert_eq!(file.segments.len(), 2);
+}
