+++ tokio-tdms/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "tokio-tdms"
+description = "NI TDMS support for Rust with Tokio"
+version = "0.3.0"
+edition = "2021"
+license = "MIT"
+repository = "https://github.com/DnOberon/tdms"
+# no need to include the test data directory in the crate, in fact that makes it so we can't publish
+exclude = ["/data"]
+
+[dependencies]
+tdms-format = { path = "../tdms-format" }
+indexmap = "1.8.2"
+thiserror = "1.0"
+hex = "0.4.3"
+extended = "0.1.0"
+fixed = "1.15.0"
+log = "0.4.17"
+tokio = { version = "1.28", features = ["full"], optional = true }
+tokio-stream = "0.1"
+async-stream = "0.3"
+async-trait = "0.1"
+async-recursion = "1.0"
+
+[dev-dependencies]
+pprof = { version = "0.10.0", features = [
+    "flamegraph",
+    "criterion",
+    "protobuf-codec",
+] }
+criterion = "0.3"
+
+
+[[bench]]
+name = "general_benchmarks"
+harness = false
+
+[features]
+default = ["tokio"]
+tokio = ["dep:tokio"]
+tokio-stream = ["dep:tokio"]
+async-recursion = ["dep:tokio"]
+async-stream = ["dep:tokio"]
+async-trait = ["dep:tokio"]
